requirements.txt
psycopg2-binary==2.7.5
Django==Django==2.1.1
python-gitlab==1.6.0
